requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
black==24.10.0
click==8.1.7
debugpy==1.8.8
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httptools==0.6.4
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
mako==1.3.6
markupsafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic[email]==2.9.2
pydantic-core==2.23.4
pyfiglet==1.0.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pyyaml==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
typing-extensions==4.12.2
uvicorn[standard]==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.0

# The following packages are considered to be unsafe in a requirements file:
setuptools==75.3.0
